# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CastleRock Investments — Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trades with optional filters
      parameters:
        - { name: accountId, in: query, schema: { type: ["integer","null"] } }
        - { name: assetClass, in: query, schema: { type: ["string","null"] } }
        - { name: symbol, in: query, schema: { type: ["string","null"] } }
        - { name: direction, in: query, schema: { type: ["string","null"] } }
        - { name: strategy, in: query, schema: { type: ["string","null"] } }
        - { name: tag, in: query, schema: { type: ["string","null"] } }
        - { name: dateFrom, in: query, schema: { type: ["string","null"] } }
        - { name: dateTo, in: query, schema: { type: ["string","null"] } }
        - { name: untaggedOnly, in: query, schema: { type: ["boolean","null"] } }
        - { name: page, in: query, schema: { type: ["integer","null"] } }
        - { name: pageSize, in: query, schema: { type: ["integer","null"] } }
        - { name: sortBy, in: query, schema: { type: ["string","null"] } }
        - { name: sortDir, in: query, schema: { type: ["string","null"] } }
      responses:
        "200":
          description: Paginated trade list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeListResponse"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Trade with journal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeDetail"
    patch:
      operationId: updateTrade
      tags: [trades]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
    delete:
      operationId: deleteTrade
      tags: [trades]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # ── Journals ──────────────────────────────────────────────────────────────
...
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      responses:
        "200":
          description: Strategy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
    post:
      operationId: createStrategy
      tags: [strategies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Created strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Strategy with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyDetail"
    patch:
      operationId: updateStrategy
      tags: [strategies]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyUpdate"
      responses:
        "200":
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
    delete:
      operationId: deleteStrategy
      tags: [strategies]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /strategies/{id}/playbook:
    get:
      operationId: getStrategyPlaybook
      tags: [strategies]
      summary: Strategy playbook export (full stats card)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Playbook export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyPlaybook"

  # ── Share Links ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id: { type: integer }
        importBatchId: { type: ["integer","null"] }
        accountId: { type: ["integer","null"] }
        importSource: { type: ["string","null"] }
        assetClass: { type: string }
        symbol: { type: string }
        instrumentDescription: { type: ["string","null"] }
        direction: { type: string }
        entryDate: { type: string }
        exitDate: { type: ["string","null"] }
        entryPrice: { type: number }
        exitPrice: { type: ["number","null"] }
        quantity: { type: number }
        pnl: { type: ["number","null"] }
        fees: { type: ["number","null"] }
        stopLoss: { type: ["number","null"] }
        takeProfit: { type: ["number","null"] }
        rMultiple: { type: ["number","null"] }
        tags: { type: ["string","null"] }
        setup: { type: ["string","null"] }
        notes: { type: ["string","null"] }
        session: { type: ["string","null"] }
        economicEventNearby: { type: ["boolean","null"] }
        hasJournal: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, assetClass, symbol, direction, entryDate, entryPrice, quantity, hasJournal, createdAt, updatedAt]
    TradeDetail:
      type: object
      properties:
        trade:
          $ref: "#/components/schemas/Trade"
        journal:
          $ref: "#/components/schemas/Journal"
      required: [trade]
    TradeUpdate:
      type: object
      properties:
        setup: { type: ["string","null"] }
        tags: { type: ["string","null"] }
        notes: { type: ["string","null"] }
        stopLoss: { type: ["number","null"] }
        takeProfit: { type: ["number","null"] }
        assetClass: { type: ["string","null"] }
        symbol: { type: ["string","null"] }
        direction: { type: ["string","null"] }
    TradeListResponse:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
      required: [trades, total, page, pageSize]

    # ── Journal ───────────────────────────────────────────────────────────────
    Journal:
      type: object
      properties:
        id: { type: integer }
        tradeId: { type: integer }
        whyEntry: { type: ["string","null"] }
        whyExit: { type: ["string","null"] }
        whyStopLoss: { type: ["string","null"] }
        whyTakeProfit: { type: ["string","null"] }
        mistakes: { type: ["string","null"] }
        marketObservation: { type: ["string","null"] }
        confidenceRating: { type: ["integer","null"] }
        ruleFollowed: { type: ["boolean","null"] }
        tiltState: { type: ["string","null"] }
        executionQualityEntry: { type: ["integer","null"] }
        executionQualityExit: { type: ["integer","null"] }
        executionQualityStop: { type: ["integer","null"] }
        strategyRulesChecked: { type: ["string","null"] }
        updatedAt: { type: string }
      required: [id, tradeId, updatedAt]
...
    AnalyticsOverview:
      type: object
      properties:
        totalTrades: { type: integer }
        winRate: { type: number }
        lossRate: { type: number }
        avgWin: { type: number }
        avgLoss: { type: number }
        expectancy: { type: number }
        expectancyR: { type: number }
        profitFactor: { type: number }
        totalPnl: { type: number }
        totalFees: { type: number }
        maxDrawdown: { type: number }
        avgRMultiple: { type: number }
        castlerockScore: { type: number }
        castlerockScoreLabel: { type: string }
        ruleAdherenceRate: { type: ["number","null"] }
        avgConfidence: { type: ["number","null"] }
        avgHoldTimeHours: { type: ["number","null"] }
        biggestWinner: { type: number }
        biggestLoser: { type: number }
        longWinRate: { type: ["number","null"] }
        shortWinRate: { type: ["number","null"] }
      required: [totalTrades, winRate, lossRate, avgWin, avgLoss, expectancy, expectancyR, profitFactor, totalPnl, totalFees, maxDrawdown, avgRMultiple, castlerockScore, castlerockScoreLabel, biggestWinner, biggestLoser]
    SliceStats:
      type: object
      properties:
        label: { type: string }
        tradeCount: { type: integer }
        winRate: { type: number }
        avgPnl: { type: number }
        totalPnl: { type: number }
        profitFactor: { type: number }
        expectancy: { type: number }
        avgRMultiple: { type: ["number","null"] }
      required: [label, tradeCount, winRate, avgPnl, totalPnl, profitFactor, expectancy]
...
    Strategy:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string","null"] }
        rules:
          type: array
          items: { type: string }
        assetClass: { type: ["string","null"] }
        createdAt: { type: string }
      required: [id, name, rules, createdAt]
    StrategyInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: ["string","null"] }
        rules:
          type: array
          items: { type: string }
        assetClass: { type: ["string","null"] }
      required: [name, rules]
    StrategyUpdate:
      type: object
      properties:
        name: { type: ["string","null"] }
        description: { type: ["string","null"] }
        rules:
          type: ["array","null"]
          items: { type: string }
    StrategyDetail:
      type: object
      properties:
        strategy:
          $ref: "#/components/schemas/Strategy"
        stats:
          $ref: "#/components/schemas/SliceStats"
        tradeCount: { type: integer }
        topSymbols:
          type: array
          items: { type: string }
      required: [strategy, tradeCount]
    StrategyPlaybook:
      type: object
      properties:
        strategy:
          $ref: "#/components/schemas/Strategy"
        stats:
          $ref: "#/components/schemas/AnalyticsOverview"
        topTrades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        generatedAt: { type: string }
      required: [strategy, stats, topTrades, generatedAt]

    # ── Share ──────────────────────────────────────────────────────────────────
...