# lib/api-zod/src/generated/types/strategyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CastleRock Investments — Trading Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface StrategyInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  rules: string[];
  /** @nullable */
  assetClass?: string | null;
}